requirements.txt
@"
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0
email-validator==2.1.0
requests==2.31.0
pandas==2.2.0
scikit-learn==1.4